+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/screens/Server/Control.tsx
@@ -1,0 +1,61 @@
+import React from 'react';
+import {
+  Collapsible,
+  Text,
+  Element,
+  Button,
+  Stack,
+} from '@codesandbox/components';
+import { useOvermind } from 'app/overmind';
+
+const RestartServerIcon = props => (
+  <svg width={12} height={11} fill="none" viewBox="0 0 12 11" {...props}>
+    <path
+      fill="#fff"
+      fillRule="evenodd"
+      d="M6.458 9.041a3.875 3.875 0 10-3.832-4.448h1.249L1.937 7.176l-.44-.588-.005.006a4.659 4.659 0 01-.006-.02L0 4.593h1.323a5.167 5.167 0 111.85 4.562l.92-.92a3.858 3.858 0 002.365.806z"
+      clipRule="evenodd"
+    />
+    <path
+      fill="#fff"
+      d="M8.22 4.934a.23.23 0 01-.012.381L5.85 6.782a.23.23 0 01-.35-.194V3.444a.23.23 0 01.362-.187l2.357 1.677z"
+    />
+  </svg>
+);
+export const Control = () => {
+  const {
+    actions: {
+      server: { restartContainer, restartSandbox },
+    },
+    state: {
+      server: { containerStatus, status },
+    },
+  } = useOvermind();
+  const disconnected = status !== 'connected';
+
+  return (
+    <Collapsible defaultOpen title="Control Container">
+      <Element paddingX={2}>
+        <Button
+          variant="secondary"
+          disabled={disconnected || containerStatus !== 'sandbox-started'}
+          onClick={restartSandbox}
+        >
+          <Stack gap={2} align="center">
+            <RestartServerIcon /> <Text>Restart Sandbox</Text>
+          </Stack>
+        </Button>
+        <Button
+          marginTop={2}
+          variant="secondary"
+          disabled={disconnected || containerStatus === 'initializing'}
+          onClick={restartContainer}
+        >
+          <Stack gap={2} align="center">
+            <RestartServerIcon /> <Text>Restart Server</Text>
+          </Stack>
+        </Button>
+      </Element>
+    </Collapsible>
+  );
+};
