+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/screens/Server/Tasks.tsx
@@ -5,15 +5,8 @@ import BuildIcon from 'react-icons/lib/fa/wrench';
 import FlaskIcon from 'react-icons/lib/fa/flask';
 import DownloadIcon from 'react-icons/lib/fa/download';
 import NodeIcon from 'react-icons/lib/io/social-nodejs';
-import { List, ListAction, Text } from '@codesandbox/components';
-
-type Props = {
-  package?: {
-    scripts?: {
-      [command: string]: string;
-    };
-  };
-};
+import { useOvermind } from 'app/overmind';
+import { List, ListAction, Text, Collapsible } from '@codesandbox/components';
 
 // These scripts are only supposed to run on the main thread.
 const blacklistedScripts = ['dev', 'develop', 'serve', 'start'];
@@ -33,7 +26,13 @@ const getIcon = (scriptName: string) => {
   }
 };
 
-export const Tasks = ({ package: pkg }: Props) => {
+export const Tasks = () => {
+  const {
+    state: {
+      editor: { parsedConfigurations: pkg },
+    },
+  } = useOvermind();
+
   const runTask = (task: string) => {
     dispatch({
       type: 'codesandbox:create-shell',
@@ -47,19 +46,22 @@ export const Tasks = ({ package: pkg }: Props) => {
   const commands = Object.keys(pkg.scripts).filter(
     x => !blacklistedScripts.includes(x)
   );
+
   return (
-    <List>
-      {commands.map(task => (
-        <ListAction onClick={() => runTask(task)} key={task}>
-          {getIcon(task)} <Text marginLeft={2}>yarn {task}</Text>
+    <Collapsible title="Run Script" defaultOpen>
+      <List>
+        {commands.map(task => (
+          <ListAction onClick={() => runTask(task)} key={task}>
+            {getIcon(task)} <Text marginLeft={2}>yarn {task}</Text>
+          </ListAction>
+        ))}
+        <ListAction onClick={() => runTask('install')}>
+          {getIcon('install')} <Text marginLeft={2}>yarn install</Text>
         </ListAction>
-      ))}
-      <ListAction onClick={() => runTask('install')}>
-        {getIcon('install')} <Text marginLeft={2}>yarn install</Text>
-      </ListAction>
-      <ListAction onClick={() => runTask('node')}>
-        {getIcon('node')} <Text marginLeft={2}>node</Text>
-      </ListAction>
-    </List>
+        <ListAction onClick={() => runTask('node')}>
+          {getIcon('node')} <Text marginLeft={2}>node</Text>
+        </ListAction>
+      </List>
+    </Collapsible>
   );
 };
